Possible fix for infinity speed

+++ labs/lab-3/sensors.ino
@@ -5,7 +5,7 @@ SensorData sensorData;
 SpeedSensor speedSensor;
 
 void setupSensorData() {
-  sensorData.speed = 0;
+  sensorData.speed = 0.0f;
 }
 
 void setupSpeedSensor(unsigned int pin) {
@@ -21,7 +21,7 @@ void speedSensorISR() {
   speedSensor.pulses++;
   if (speedSensor.pulses % 4) {
     speedSensor.t0 = speedSensor.t1;
-    speedSensor.t1 = micros()/1000000;
+    speedSensor.t1 = (int) (micros()/1000000);
     setSpeed(CALC_SPEED(speedSensor.t0, speedSensor.t1, (WHEEL_DIAM_CM/2)*100));
   }
 }
